improvement: push everything under one `blocktorch` pkg
