+++ proto/combined_rollups.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package eth.rollup.v1;
+package eth.blocktorch.v1;
 
 import "nitro_rollup.proto";
 import "op_rollup.proto";
