+++ proto/combined_transactions.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package eth.transaction.v1;
+package eth.blocktorch.v1;
 
 import "aa_transaction.proto";
 import "chronicle_transaction.proto";
