+++ proto/zk_evm_rollup.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package eth.rollup.v1;
+package eth.blocktorch.v1;
 
 import "google/protobuf/timestamp.proto";
