+++ src/map_filter_aa_transactions.rs
@@ -1,4 +1,4 @@
-use crate::pb::eth::transaction::v1::{AccountAbstractionTransaction as Transaction, AccountAbstractionTransactions as Transactions};
+use crate::pb::eth::blocktorch::v1::{AccountAbstractionTransaction as Transaction, AccountAbstractionTransactions as Transactions};
 use crate::abi;
 use substreams::{log, Hex};
 use substreams_ethereum::{Function, Event};
